+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Current develop
 
-### Fixed cmake code to properly install fortran module files in include path
+### Fixed (Repair bugs, etc)
+- [[PR157]](https://github.com/lanl/singularity-eos/pull/157) fix root finders in Gruneisen, Davis, and JWL
 - [[PR151]](https://github.com/lanl/singularity-eos/pull/151) fix module install
 
 ### Added (new features/APIs/variables/...)